Code cleanup and documentation
