+++ src/SmartOffice.Models/PartnerCenter/Address.cs
@@ -8,23 +8,54 @@ namespace Microsoft.Partner.SmartOffice.Models.PartnerCenter
 {
     public sealed class Address
     {
+        /// <summary>
+        /// Gets or sets the first address line.
+        /// </summary>
         public string AddressLine1 { get; set; }
 
+        /// <summary>
+        /// Gets or sets the second address line.
+        /// </summary>
         public string AddressLine2 { get; set; }
 
+        /// <summary>
+        /// Gets or sets the city.
+        /// </summary>
         public string City { get; set; }
 
+        /// <summary>
+        /// Gets or sets the country.
+        /// </summary>
         public string Country { get; set; }
 
+        /// <summary>
+        /// Gets or sets the first name.
+        /// </summary>
         public string FirstName { get; set; }
 
+        /// <summary>
+        /// Gets or sets the last name.
+        /// </summary>
         public string LastName { get; set; }
+
+        /// <summary>
+        /// Gets or sets the phone number.
+        /// </summary>
         public string PhoneNumber { get; set; }
 
+        /// <summary>
+        /// Gets or sets the postal code.
+        /// </summary>
         public string PostalCode { get; set; }
 
+        /// <summary>
+        /// Gets or sets the region.
+        /// </summary>
         public string Region { get; set; }
 
+        /// <summary>
+        /// Gets or sets the state.
+        /// </summary>
         public string State { get; set; }
     }
 }
