+++ src/SmartOffice.Models/PartnerCenter/Offers/Product.cs
@@ -8,10 +8,19 @@ namespace Microsoft.Partner.SmartOffice.Models.PartnerCenter.Offers
 {
     public sealed class Product
     {
+        /// <summary>
+        /// Gets or sets the identifier of the product.
+        /// </summary>
         public string Id { get; set; }
 
+        /// <summary>
+        /// Gets or sets the name of the product.
+        /// </summary>
         public string Name { get; set; }
 
+        /// <summary>
+        /// Gets or sets the product unti.
+        /// </summary>
         public string Unit { get; set; }
     }
 }
