+++ src/SmartOffice.Models/PartnerCenter/ResourceCollectionWithLinks.cs
@@ -1,5 +1,5 @@
 ﻿// -----------------------------------------------------------------------
-// <copyright file="ResourceCollection.cs" company="Microsoft">
+// <copyright file="ResourceCollectionWithLinks.cs" company="Microsoft">
 //     Copyright (c) Microsoft Corporation. All rights reserved.
 // </copyright>
 // -----------------------------------------------------------------------
@@ -16,12 +16,20 @@ namespace Microsoft.Partner.SmartOffice.Models
     {
         private readonly ICollection<TResource> internalItems;
 
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ResourceCollection{TResource, TLinks}" /> class.
+        /// </summary>
+        /// <param name="items">The collection whose elements are copied to the new list.</param>
         public ResourceCollection(ICollection<TResource> items)
           : base("Collection")
         {
             internalItems = items ?? new List<TResource>();
         }
 
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ResourceCollection{TResource, TLinks}" /> class.
+        /// </summary>
+        /// <param name="resourceCollection">The resource collection whose elements are copied to the new list.</param>
         protected ResourceCollection(ResourceCollection<TResource, TLinks> resourceCollection)
           : base("Collection")
         {
@@ -33,8 +41,14 @@ namespace Microsoft.Partner.SmartOffice.Models
             internalItems = resourceCollection.internalItems;
         }
 
+        /// <summary>
+        /// Gets the items of the resource collection.
+        /// </summary>
         public IEnumerable<TResource> Items => internalItems;
 
+        /// <summary>
+        /// Gets the total count of the elements in the resource collection.
+        /// </summary>
         [JsonProperty]
         public int TotalCount => internalItems.Count;
     }
