+++ src/SmartOffice.Models/PartnerCenter/Subscriptions/BillingType.cs
@@ -4,7 +4,7 @@
 // </copyright>
 // -----------------------------------------------------------------------
 
-namespace Microsoft.Partner.SmartOffice.Models.PartnerCenter
+namespace Microsoft.Partner.SmartOffice.Models.PartnerCenter.Subscriptions
 {
     public enum BillingType
     {
