+++ src/SmartOffice.Models/PartnerCenter/Subscriptions/SubscriptionStatus.cs
@@ -6,13 +6,9 @@
 
 namespace Microsoft.Partner.SmartOffice.Models.PartnerCenter.Subscriptions
 {
-    using Converters;
-    using Newtonsoft.Json;
-
     /// <summary>
     /// Lists the available states for a subscription.
     /// </summary>
-    [JsonConverter(typeof(EnumJsonConverter))]
     public enum SubscriptionStatus
     {
         /// <summary>
