+++ src/SmartOffice.Services/PartnerCenter/AuditRecords/AuditRecordCollectionOperations.cs
@@ -11,6 +11,7 @@ namespace Microsoft.Partner.SmartOffice.Services.PartnerCenter.AuditRecords
     using System.Threading;
     using System.Threading.Tasks;
     using Models.PartnerCenter;
+    using Models.PartnerCenter.AuditRecords;
 
     public class AuditRecordCollectionOperations : IAuditRecordCollectionOperations
     {
@@ -28,6 +29,12 @@ namespace Microsoft.Partner.SmartOffice.Services.PartnerCenter.AuditRecords
             this.client = client;
         }
 
+        /// <summary>
+        /// Queries audit records associated to the partner.
+        /// </summary>
+        /// <param name="nextLink">An instance of the <see cref="Link" /> class that represents the next page to be requested.</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>The audit records that match the given query.</returns>
         public async Task<SeekBasedResourceCollection<AuditRecord>> QueryAsync(Link nextLink, CancellationToken cancellationToken = default(CancellationToken))
         {
             return await client.GetAsync<SeekBasedResourceCollection<AuditRecord>>(
@@ -35,6 +42,13 @@ namespace Microsoft.Partner.SmartOffice.Services.PartnerCenter.AuditRecords
                 cancellationToken).ConfigureAwait(false);
         }
 
+        /// <summary>
+        /// Queries audit records associated to the partner.
+        /// </summary>
+        /// <param name="startDate">The start date of the audit record logs.</param>
+        /// <param name="endDate">The end date of the audit record logs</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>The audit records that match the given query.</returns>
         public async Task<SeekBasedResourceCollection<AuditRecord>> QueryAsync(DateTime startDate, DateTime? endDate = null, CancellationToken cancellationToken = default(CancellationToken))
         {
             return await client.GetAsync<SeekBasedResourceCollection<AuditRecord>>(
